requirements.txt
aiohttp
asyncio
pymongo
google-generativeai
fastapi
uvicorn
python-dotenv
pytest
httpx

